With new ffmpeg libraries
Functionally the same but built with the most recent ffmpeg libraries. So if you aren't having any issues then you may not want to bother to update.
We have a 64 bit andriod version now for the Google Pixel 8s and above.
On the firestick the CPU usage looks better and I bumped up the buffering by 200mSec hopefully this will help some peoples reported stuttering.